🏗️ Rename InstallPWAButton to AddPWAButton and refactor its implementation
